Add API to turn a Namespace package into a series of "add" changes
